修正 ls 实现cd

+++ tools/LINUX_TOOLS/1.4/temp_src/stringio_h.h
@@ -4,7 +4,7 @@ extern void printInPos(char*msg,uint16_t len,uint8_t row,uint8_t col);
 extern void putchar(char c);         // Output one character
 extern char getch();                // Get keyboard input
 
-uint16_t strlen(char *str);
+uint32_t strlen(char *str);
 void print(char *s);
 void printint(uint32_t num);
 void readBuf(char * buffer , uint32_t maxlen);
